BF: Custom argparse formatters CANNOT be imported!
https://travis-ci.org/datalad/datalad/jobs/136515428
